+++ lambda/chalicelib/util/ccavutil.py
@@ -3,39 +3,6 @@ import hashlib
 from urllib.parse import urlencode, parse_qsl
 import binascii
 
-POSSIBLE_PARAMS = [
-    "merchant_id",
-    "order_id",
-    "currency",
-    "amount",
-    "redirect_url",
-    "cancel_url",
-    "language",
-    "billing_name",
-    "billing_address",
-    "billing_city",
-    "billing_state",
-    "billing_zip",
-    "billing_country",
-    "billing_tel",
-    "billing_email",
-    "delivery_name",
-    "delivery_address",
-    "delivery_city",
-    "delivery_state",
-    "delivery_zip",
-    "delivery_country",
-    "delivery_tel",
-    "merchant_param1",
-    "merchant_param2",
-    "merchant_param3",
-    "merchant_param4",
-    "merchant_param5",
-    "integration_type",
-    "promo_code",
-    "customer_identifier",
-]
-
 
 def pad(data):
     length = 16 - (len(data) % 16)