+++ lambda/chalicelib/util/formSubmit/ccavenue.py
@@ -51,6 +51,9 @@ def update_ccavenue_hash(formId, ccavenuePaymentMethodInfo, response):
             ),
         ),
     }
+    sub_account_id = ccavenuePaymentMethodInfo.get("sub_account_id")
+    if sub_account_id is not None:
+        data["sub_account_id"] = sub_account_id
     # "delivery_name": "test",
     # "delivery_address": "test",
     # "delivery_city": "test",